microservices/storer/requirements.txt
grpcio==1.34.1
grpcio-tools==1.34.1
protobuf==3.14.0
six==1.15.0
toml~=0.10.2
influxdb~=5.3.1
